README.md
Idea:
Trying to save money always seems like a chore and hard to be consistent in doing so. What if we can gamify the process of saving money. What if we can make a virtual pet that can motivate you to save more and spend less.
This is where our idea comes in.


The way our idea works is the user gets prompted to enter an amount they want to save the next month or year. The amount they enter will be used as a health bar for their pet. The user can select which critter they want as a pet. In the following days the transactions made by the user will either damage the pet(spending money) or heal it(earning money). The pet will be at max health/hp bar will be full once the saving goal gets reached. Every transaction will let the user earn points which can be used to spend in a shop. You can use the points to roll accessories for your pets. You can do more transactions to get points but you have to be careful if you go on a spending spree the pet will get hurt. We can't have our cute critters die due to overspending now can we?

Problem:
To reduce spendings and save money

How our idea solves it:
Making a sentimental attachment of the user to the pet can help them save money.

Tech stack: